Fix request mapping and file system write
