+++ build.gradle
@@ -91,10 +91,6 @@ dependencies {
 	compile project(":rules")
     compile project(":filesystem")
 
-	compile group: 'net.lingala.zip4j', name: 'zip4j', version: '1.3.2'
-
-	compile group: 'net.harawata', name: 'appdirs', version: '1.0.1'
-
 	compile 'com.github.BilledTrain380:Alberta:447bf3a091f8128b4a6122c7c9b6c4508527f509'
 	
     // database dependencies
