+++ filesystem/src/main/kotlin/ch/schulealtendorf/sporttagpsa/filesystem/PlatformFileSystem.kt
@@ -106,8 +106,8 @@ class PlatformFileSystem(
                 }
 
             } finally {
-                it.close()
                 it.flush()
+                it.close()
             }
         }
