--- src/main/kotlin/ch/schulealtendorf/sporttagpsa/business/export/archive/ZipManager.kt
@@ -1,81 +1,0 @@
-/*
- * Copyright (c) 2018 by Nicolas Märchy
- *
- * This file is part of Sporttag PSA.
- *
- * Sporttag PSA is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * Sporttag PSA is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with Sporttag PSA.  If not, see <http://www.gnu.org/licenses/>.
- *
- * Diese Datei ist Teil von Sporttag PSA.
- *
- * Sporttag PSA ist Freie Software: Sie können es unter den Bedingungen
- * der GNU General Public License, wie von der Free Software Foundation,
- * Version 3 der Lizenz oder (nach Ihrer Wahl) jeder späteren
- * veröffentlichten Version, weiterverbreiten und/oder modifizieren.
- *
- * Sporttag PSA wird in der Hoffnung, dass es nützlich sein wird, aber
- * OHNE JEDE GEWÄHRLEISTUNG, bereitgestellt; sogar ohne die implizite
- * Gewährleistung der MARKTFÄHIGKEIT oder EIGNUNG FÜR EINEN BESTIMMTEN ZWECK.
- * Siehe die GNU General Public License für weitere Details.
- *
- * Sie sollten eine Kopie der GNU General Public License zusammen mit diesem
- * Programm erhalten haben. Wenn nicht, siehe <http://www.gnu.org/licenses/>.
- *
- *
- */
-
-package ch.schulealtendorf.sporttagpsa.business.export.archive
-
-import ch.schulealtendorf.sporttagpsa.business.storage.StorageManager
-import net.lingala.zip4j.core.ZipFile
-import net.lingala.zip4j.model.ZipParameters
-import net.lingala.zip4j.util.Zip4jConstants
-import org.springframework.stereotype.Component
-import java.io.File
-
-/**
- * Archive manager which manages .zip files.
- * 
- * @author nmaerchy
- * @version 1.0.0
- */
-@Component
-class ZipManager(
-        private val storageManager: StorageManager
-): ArchiveManager {
-
-    /**
-     * Creates an archive with the given {@code files}.
-     * This method considers the platform depended special directory.
-     *
-     * @param outputFileName the file name of the created archive without the extension
-     * @param files files to include in the archive
-     *
-     * @return the generated archive
-     */
-    override fun createArchive(outputFileName: String, files: Iterable<File>): File {
-        
-        val zipFile = storageManager.getApplicationDir().resolve("$outputFileName.zip")
-        zipFile.delete()
-
-        val rankingZip = ZipFile(zipFile)
-        val parameters = ZipParameters().apply {
-            compressionMethod = Zip4jConstants.COMP_DEFLATE
-            compressionLevel = Zip4jConstants.DEFLATE_LEVEL_NORMAL
-        }
-
-        files.forEach { rankingZip.addFile(it, parameters) }
-
-        return zipFile
-    }
-}
