--- src/main/resources/templates/competitor/management-basic.html
@@ -1,105 +1,0 @@
-<!DOCTYPE html>
-<!--
-  ~ Copyright (c) 2017 by Nicolas Märchy
-  ~
-  ~ This file is part of Sporttag PSA.
-  ~
-  ~ Sporttag PSA is free software: you can redistribute it and/or modify
-  ~ it under the terms of the GNU General Public License as published by
-  ~ the Free Software Foundation, either version 3 of the License, or
-  ~ (at your option) any later version.
-  ~
-  ~ Sporttag PSA is distributed in the hope that it will be useful,
-  ~ but WITHOUT ANY WARRANTY; without even the implied warranty of
-  ~ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-  ~ GNU General Public License for more details.
-  ~
-  ~ You should have received a copy of the GNU General Public License
-  ~ along with Sporttag PSA.  If not, see <http://www.gnu.org/licenses/>.
-  ~
-  ~ Diese Datei ist Teil von Sporttag PSA.
-  ~
-  ~ Sporttag PSA ist Freie Software: Sie können es unter den Bedingungen
-  ~ der GNU General Public License, wie von der Free Software Foundation,
-  ~ Version 3 der Lizenz oder (nach Ihrer Wahl) jeder späteren
-  ~ veröffentlichten Version, weiterverbreiten und/oder modifizieren.
-  ~
-  ~ Sporttag PSA wird in der Hoffnung, dass es nützlich sein wird, aber
-  ~ OHNE JEDE GEWÄHRLEISTUNG, bereitgestellt; sogar ohne die implizite
-  ~ Gewährleistung der MARKTFÄHIGKEIT oder EIGNUNG FÜR EINEN BESTIMMTEN ZWECK.
-  ~ Siehe die GNU General Public License für weitere Details.
-  ~
-  ~ Sie sollten eine Kopie der GNU General Public License zusammen mit diesem
-  ~ Programm erhalten haben. Wenn nicht, siehe <http://www.gnu.org/licenses/>.
-  ~
-  ~
-  -->
-
-<html lang="de">
-
-<head th:replace="fragments/main :: head(Verwaltung)">
-</head>
-
-<body id="page-top">
-
-<!-- Navigation -->
-<nav th:replace="fragments/main :: navigation(competitor, management)">
-</nav>
-
-<div class="content-wrapper py-3">
-
-    <div class="container-fluid">
-
-        <div class="page-header">
-            <h1>Teilnehmer <small>Verwaltung</small></h1>
-        </div>
-
-        <div class="row">
-
-            <div class="col-lg-12">
-
-                <ul class="nav nav-tabs">
-                    <li class="nav-item">
-                        <a class="nav-link active" th:href="@{/management/basic}">Allgemein</a>
-                    </li>
-
-                    <li class="nav-item">
-                        <a class="nav-link" th:href="@{/management/participant-list}">Teilnehmerliste</a>
-                    </li>
-                </ul>
-                
-            </div>
-            
-            <div class="col-lg-12">
-                
-                <br/>
-
-                <div th:if="${participationInfo}" class="alert alert-info" role="alert">
-                    Sie haben die Teilnahme abgeschlossen. Sie können nun die Startlisten exportieren.
-                </div>
-                
-                <h4>Teilnahme
-                    <span th:if="${!participationStatus}" class="badge badge-success">offen</span>
-                    <span th:if="${participationStatus}" class="badge badge-info">abgeschlossen</span>
-                    <a th:if="${!participationStatus}" class="btn btn-primary btn-sm" th:href="@{/management/finish-participation}">Teilnahme abschliessen</a>
-                </h4>
-                <p>
-                    Schliessen Sie die Teilnahme am Sporttag ab, um die Startlisten zu generieren.
-                    Wenn Sie diese Aktion ausführen können Sie nicht länger Änderungen an der Sportart
-                    in der Klassenliste vornehmen.
-                </p>
-                
-            </div>
-
-        </div>
-
-    </div>
-    <!-- /.container-fluid -->
-
-</div>
-<!-- /.content-wrapper -->
-
-
-</body>
-
-</html>
