+++ src/main/resources/templates/fragments/main.html
@@ -100,7 +100,7 @@
                         <a th:href="@{/clazz}">Klassenliste</a>
                     </li>
                     <li th:class="${active} == management? active : ''">
-                        <a th:href="@{/management/basic}">Verwaltung</a>
+                        <a th:href="@{/participant/management/general}">Verwaltung</a>
                     </li>
                 </ul>
             </li>
