+++ src/main/resources/templates/participant/management/participant-list.html
@@ -37,7 +37,7 @@
 
 <html lang="de">
 
-<head th:replace="fragments/main :: head(Teilnehmer - Verwaltung)">
+<head th:replace="fragments/main :: head('Teilnehmer - Verwaltung')">
 </head>
 
 <body id="page-top">
@@ -56,19 +56,7 @@
 
         <div class="row">
 
-            <div class="col-lg-12">
-
-                <ul class="nav nav-tabs">
-                    <li class="nav-item">
-                        <a class="nav-link" th:href="@{/management/basic}">Allgemein</a>
-                    </li>
-
-                    <li class="nav-item">
-                        <a class="nav-link active" th:href="@{/management/participant-list}">Teilnehmerliste</a>
-                    </li>
-                </ul>
-
-            </div>
+            <div th:replace="participant/management/tab-fragment :: tab-menu(participant-list)"> </div>
 
             <div class="col-lg-12">
