+++ src/main/resources/templates/ranking/ranking.html
@@ -37,7 +37,7 @@
 
 <html lang="de">
 
-<head th:replace="fragments/main :: head(Mehrkampf)">
+<head th:replace="fragments/main :: head(Rangliste)">
 </head>
 
 <body id="page-top">
