+++ src/main/resources/templates/tournament/result.html
@@ -37,7 +37,7 @@
 
 <html lang="de">
 
-<head th:replace="fragments/main :: head(Mehrkampf - Resultate)">
+<head th:replace="fragments/main :: head('Mehrkampf - Resultate')">
 </head>
 
 <body id="page-top">
